chore: centralize processor and config protocols
