+++ services/configuration_service.py
@@ -1,29 +1,12 @@
 from __future__ import annotations
 
-from typing import Any, Dict, Protocol, runtime_checkable
+from typing import Any, Dict
+
+from core.protocols import ConfigurationServiceProtocol
 
 from config.dynamic_config import DynamicConfigManager, dynamic_config
 
 
-@runtime_checkable
-class ConfigurationServiceProtocol(Protocol):
-    """Interface for accessing runtime configuration values."""
-
-    def get_max_upload_size_mb(self) -> int: ...
-
-    def get_max_upload_size_bytes(self) -> int: ...
-
-    def validate_large_file_support(self) -> bool: ...
-
-    def get_upload_chunk_size(self) -> int: ...
-
-    def get_max_parallel_uploads(self) -> int: ...
-
-    def get_validator_rules(self) -> Dict[str, Any]: ...
-
-    def get_ai_confidence_threshold(self) -> int: ...
-
-    def get_db_pool_size(self) -> int: ...
 
 
 class DynamicConfigurationService(ConfigurationServiceProtocol):
