+++ services/door_mapping_service.py
@@ -16,10 +16,8 @@ import pandas as pd
 from services.ai_device_generator import AIDeviceGenerator
 from services.common import ModelRegistry
 from services.common.config_utils import common_init, create_config_methods
-from services.configuration_service import (
-    ConfigurationServiceProtocol,
-    DynamicConfigurationService,
-)
+from core.protocols import ConfigurationServiceProtocol
+from services.configuration_service import DynamicConfigurationService
 from services.learning.src.api.consolidated_service import get_learning_service
 
 logger = logging.getLogger(__name__)
