+++ services/file_processor_service.py
@@ -16,7 +16,7 @@ import pandas as pd
 from analytics_core.utils.unicode_processor import UnicodeHelper
 from config.constants import UPLOAD_ALLOWED_EXTENSIONS
 from core.unicode import process_large_csv_content
-from services.configuration_service import ConfigurationServiceProtocol
+from core.protocols import ConfigurationServiceProtocol
 from utils.file_utils import safe_decode_with_unicode_handling
 from utils.memory_utils import memory_safe
 from utils.protocols import SafeDecoderProtocol
