+++ services/upload/service_registration.py
@@ -23,12 +23,12 @@ from services.upload.core.file_processor_service import FileProcessor
 from services.upload.core.learning_coordinator import LearningCoordinator
 from services.upload.core.ui_builder import UploadUIBuilder
 from services.upload.protocols import (
-    FileProcessorProtocol,
     UploadControllerProtocol,
     UploadProcessingServiceProtocol,
     UploadStorageProtocol,
     UploadValidatorProtocol,
 )
+from core.protocols import FileProcessorProtocol
 from services.upload_data_service import UploadDataService
 from utils.upload_store import UploadedDataStore
